+++ src/api/Migrations/20250604093951_AddDocument.cs
@@ -1,0 +1,93 @@
+﻿using Microsoft.EntityFrameworkCore.Migrations;
+using Npgsql.EntityFrameworkCore.PostgreSQL.Metadata;
+
+#nullable disable
+
+namespace BDMS.Migrations;
+
+/// <inheritdoc />
+public partial class AddDocument : Migration
+{
+    /// <inheritdoc />
+    protected override void Up(MigrationBuilder migrationBuilder)
+    {
+        migrationBuilder.AddColumn<bool>(
+            name: "document",
+            schema: "bdms",
+            table: "tab_status",
+            type: "boolean",
+            nullable: false,
+            defaultValue: false);
+
+        migrationBuilder.CreateTable(
+            name: "document",
+            schema: "bdms",
+            columns: table => new
+            {
+                id = table.Column<int>(type: "integer", nullable: false)
+                    .Annotation("Npgsql:ValueGenerationStrategy", NpgsqlValueGenerationStrategy.IdentityByDefaultColumn),
+                borehole_id = table.Column<int>(type: "integer", nullable: false),
+                url = table.Column<string>(type: "text", nullable: false),
+                description = table.Column<string>(type: "text", nullable: true),
+                @public = table.Column<bool>(name: "public", type: "boolean", nullable: false),
+                creator = table.Column<int>(type: "integer", nullable: true),
+                creation = table.Column<DateTime>(type: "timestamp with time zone", nullable: true),
+                updater = table.Column<int>(type: "integer", nullable: true),
+                update = table.Column<DateTime>(type: "timestamp with time zone", nullable: true),
+            },
+            constraints: table =>
+            {
+                table.PrimaryKey("PK_document", x => x.id);
+                table.ForeignKey(
+                    name: "FK_document_borehole_borehole_id",
+                    column: x => x.borehole_id,
+                    principalSchema: "bdms",
+                    principalTable: "borehole",
+                    principalColumn: "id_bho",
+                    onDelete: ReferentialAction.Cascade);
+                table.ForeignKey(
+                    name: "FK_document_users_creator",
+                    column: x => x.creator,
+                    principalSchema: "bdms",
+                    principalTable: "users",
+                    principalColumn: "id_usr");
+                table.ForeignKey(
+                    name: "FK_document_users_updater",
+                    column: x => x.updater,
+                    principalSchema: "bdms",
+                    principalTable: "users",
+                    principalColumn: "id_usr");
+            });
+
+        migrationBuilder.CreateIndex(
+            name: "IX_document_borehole_id",
+            schema: "bdms",
+            table: "document",
+            column: "borehole_id");
+
+        migrationBuilder.CreateIndex(
+            name: "IX_document_creator",
+            schema: "bdms",
+            table: "document",
+            column: "creator");
+
+        migrationBuilder.CreateIndex(
+            name: "IX_document_updater",
+            schema: "bdms",
+            table: "document",
+            column: "updater");
+    }
+
+    /// <inheritdoc />
+    protected override void Down(MigrationBuilder migrationBuilder)
+    {
+        migrationBuilder.DropTable(
+            name: "document",
+            schema: "bdms");
+
+        migrationBuilder.DropColumn(
+            name: "document",
+            schema: "bdms",
+            table: "tab_status");
+    }
+}